Splitting the EqualsNode into two parts
